Grouped message entities
