+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "botx"
-version = "0.15.17"
+version = "0.15.18"
 description = "A little python framework for building bots for eXpress"
 license = "MIT"
 authors = ["Sidnev Nikolay <nsidnev@ccsteam.ru>"]
